requirements.txt
# external dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.28.1
pinecone==7.0.2

# langchain
langchain==0.3.25
langchain-core>=0.3.58,<1.0.0
langchain-community==0.3.12
